feat(ai): move the machine index to /ai/home
/ai/home is now the machine view's landing page and first nav entry
(inverted-highlighted when active), replacing the bare / human link.
/ai 308-redirects to it; the toggle, footers, sitemap, llms.txt and
agents.md all point at the new path.
