+++ src/app/ai/blog/[[...slug]]/page.tsx
@@ -115,7 +115,7 @@ const MachineBlogCategoryPage = async ({
 
         <footer className="flex flex-col gap-1 text-machine-dim">
           <p>
-            <a href="/ai" className={linkClass}>
+            <a href="/ai/home" className={linkClass}>
               back to machine index
             </a>{" "}
             ·{" "}
@@ -238,7 +238,7 @@ const MachineBlogIndexPage = async () => {
 
         <footer className="flex flex-col gap-1 text-machine-dim">
           <p>
-            <a href="/ai" className={linkClass}>
+            <a href="/ai/home" className={linkClass}>
               back to machine index
             </a>{" "}
             ·{" "}
