+++ src/app/ai/careers/[slug]/page.tsx
@@ -104,7 +104,7 @@ const MachineCareerPage = async ({ params }: MachineCareerProps) => {
 
         <footer className="flex flex-col gap-1 uppercase text-machine-dim">
           <p>
-            <a href="/ai" className={linkClass}>
+            <a href="/ai/home" className={linkClass}>
               back to machine index
             </a>{" "}
             ·{" "}
