+++ src/app/ai/contact/page.tsx
@@ -102,7 +102,7 @@ const MachineContactPage = async () => {
 
         <footer className="flex flex-col gap-1 text-machine-dim">
           <p>
-            <a href="/ai" className={linkClass}>
+            <a href="/ai/home" className={linkClass}>
               back to machine index
             </a>{" "}
             ·{" "}
