+++ src/app/ai/people/page.tsx
@@ -168,7 +168,7 @@ const MachinePeoplePage = async () => {
 
         <footer className="flex flex-col gap-1 uppercase text-machine-dim">
           <p>
-            <a href="/ai" className={linkClass}>
+            <a href="/ai/home" className={linkClass}>
               back to machine index
             </a>{" "}
             ·{" "}
