+++ src/app/ai/post/[slug]/page.tsx
@@ -110,7 +110,7 @@ const MachinePostPage = async ({ params }: MachinePostProps) => {
 
         <footer className="flex flex-col gap-1 uppercase text-machine-dim">
           <p>
-            <a href="/ai" className={linkClass}>
+            <a href="/ai/home" className={linkClass}>
               back to machine index
             </a>{" "}
             ·{" "}
