+++ src/app/ai/services/page.tsx
@@ -155,7 +155,7 @@ const MachineServicesPage = async () => {
 
         <footer className="flex flex-col gap-1 uppercase text-machine-dim">
           <p>
-            <a href="/ai" className={linkClass}>
+            <a href="/ai/home" className={linkClass}>
               back to machine index
             </a>{" "}
             ·{" "}
