+++ src/app/ai/showcase/[slug]/page.tsx
@@ -157,7 +157,7 @@ const MachineProjectPage = async ({ params }: MachineProjectProps) => {
 
         <footer className="flex flex-col gap-1 uppercase text-machine-dim">
           <p>
-            <a href="/ai" className={linkClass}>
+            <a href="/ai/home" className={linkClass}>
               back to machine index
             </a>{" "}
             ·{" "}
