+++ src/app/ai/showcase/page.tsx
@@ -83,7 +83,7 @@ const MachineShowcasePage = async () => {
 
         <footer className="flex flex-col gap-1 text-machine-dim">
           <p>
-            <a href="/ai" className={linkClass}>
+            <a href="/ai/home" className={linkClass}>
               back to machine index
             </a>{" "}
             ·{" "}
