+++ src/app/sitemap.ts
@@ -42,7 +42,7 @@ const staticRoutes: Array<{ href: string; priority: number }> = [
   { href: "/people", priority: 0.7 },
   { href: "/contact", priority: 0.7 },
   { href: "/faq", priority: 0.7 },
-  { href: "/ai", priority: 0.5 },
+  { href: "/ai/home", priority: 0.5 },
   { href: "/lab", priority: 0.5 },
   { href: "/basketball", priority: 0.3 },
   { href: "/doom", priority: 0.3 }
